Reorganize test utilities in hardware_interfaces
