+++ panther_hardware_interfaces/test/panther_system/motors_controller/test_roboteq_driver.cpp
@@ -23,8 +23,8 @@
 #include <panther_hardware_interfaces/panther_system/motors_controller/canopen_controller.hpp>
 #include <panther_hardware_interfaces/panther_system/motors_controller/roboteq_driver.hpp>
 
-#include <roboteqs_mock.hpp>
-#include <test_constants.hpp>
+#include "utils/roboteqs_mock.hpp"
+#include "utils/test_constants.hpp"
 
 #include <iostream>
