+++ panther_hardware_interfaces/test/utils/roboteqs_mock.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef PANTHER_HARDWARE_INTERFACES_ROBOTEQ_MOCK_HPP_
-#define PANTHER_HARDWARE_INTERFACES_ROBOTEQ_MOCK_HPP_
+#ifndef PANTHER_HARDWARE_INTERFACES_TEST_UTILS_ROBOTEQ_MOCK_HPP_
+#define PANTHER_HARDWARE_INTERFACES_TEST_UTILS_ROBOTEQ_MOCK_HPP_
 
 #include <atomic>
 #include <condition_variable>
@@ -83,9 +83,25 @@ public:
   {
     (*this)[0x2100][static_cast<std::uint8_t>(channel)] = value;
   }
-  void SetDriverFaultFlag(const DriverFaultFlags flag);
-  void SetDriverScriptFlag(const DriverScriptFlags flag);
-  void SetDriverRuntimeError(const DriverChannel channel, const DriverRuntimeErrors flag);
+  void SetDriverFaultFlag(const DriverFaultFlags flag)
+  {
+    std::int32_t current_data = (*this)[0x2106][7];
+    current_data |= (0b00000001 << static_cast<std::uint8_t>(flag));
+    (*this)[0x2106][7] = current_data;
+  }
+  void SetDriverScriptFlag(const DriverScriptFlags flag)
+  {
+    std::int32_t current_data = (*this)[0x2106][7];
+    current_data |= std::int32_t(0b00000001 << static_cast<std::uint8_t>(flag)) << 3 * 8;
+    (*this)[0x2106][7] = current_data;
+  }
+  void SetDriverRuntimeError(const DriverChannel channel, const DriverRuntimeErrors flag)
+  {
+    std::int32_t current_data = (*this)[0x2106][7];
+    current_data |= static_cast<std::int32_t>(0b00000001 << static_cast<std::uint8_t>(flag))
+                    << (static_cast<std::uint8_t>(channel)) * 8;
+    (*this)[0x2106][7] = current_data;
+  }
   void SetTemperature(const std::int16_t value) { (*this)[0x210F][1] = value; }
   void SetHeatsinkTemperature(const std::int16_t value) { (*this)[0x210F][2] = value; }
   void SetVoltage(const std::uint16_t value) { (*this)[0x210D][2] = value; }
@@ -117,7 +133,23 @@ public:
   /**
    * @brief Sets initial values (positions, temperatures, etc.) to zeros
    */
-  void InitializeValues();
+  void InitializeValues()
+  {
+    SetTemperature(0);
+    SetHeatsinkTemperature(0);
+    SetVoltage(0);
+    SetBatteryCurrent1(0);
+    SetBatteryCurrent2(0);
+
+    SetPosition(DriverChannel::CHANNEL1, 0);
+    SetPosition(DriverChannel::CHANNEL2, 0);
+    SetVelocity(DriverChannel::CHANNEL1, 0);
+    SetVelocity(DriverChannel::CHANNEL2, 0);
+    SetCurrent(DriverChannel::CHANNEL1, 0);
+    SetCurrent(DriverChannel::CHANNEL2, 0);
+
+    ClearErrorFlags();
+  };
 
   /**
    * @brief Creates two threads, one that will trigger motors states PDOs and second that triggers
@@ -128,12 +160,40 @@ public:
    */
   void StartPublishing(
     const std::chrono::milliseconds motors_states_period,
-    const std::chrono::milliseconds driver_state_period);
+    const std::chrono::milliseconds driver_state_period)
+  {
+    motors_states_publishing_thread_ = std::thread([this, motors_states_period]() {
+      auto next = std::chrono::steady_clock::now();
+      while (!stop_publishing_) {
+        next += motors_states_period;
+        TriggerMotorsStatesPublish();
+        std::this_thread::sleep_until(next);
+      }
+    });
+
+    driver_state_publishing_thread_ = std::thread([this, driver_state_period]() {
+      auto next = std::chrono::steady_clock::now();
+      while (!stop_publishing_) {
+        next += driver_state_period;
+        TriggerDriverStatePublish();
+        std::this_thread::sleep_until(next);
+      }
+    });
+  }
 
   /**
    * @brief Stops publishing threads
    */
-  void StopPublishing();
+  void StopPublishing()
+  {
+    stop_publishing_.store(true);
+    if (motors_states_publishing_thread_.joinable()) {
+      motors_states_publishing_thread_.join();
+    }
+    if (driver_state_publishing_thread_.joinable()) {
+      driver_state_publishing_thread_.join();
+    }
+  }
 
   /**
    * @brief Adds sleep when write event for given CANopen object was registered. Can be used for
@@ -178,8 +238,18 @@ public:
   }
 
 private:
-  void TriggerMotorsStatesPublish();
-  void TriggerDriverStatePublish();
+  void TriggerMotorsStatesPublish()
+  {
+    // Every PDO holds two values - it is enough to send an event to just one and both will be sent
+    this->WriteEvent(0x2104, 1);
+    this->WriteEvent(0x2107, 1);
+  }
+  void TriggerDriverStatePublish()
+  {
+    // Every PDO holds two values - it is enough to send an event to just one and both will be sent
+    this->WriteEvent(0x2106, 7);
+    this->WriteEvent(0x210D, 2);
+  }
 
   std::thread motors_states_publishing_thread_;
   std::thread driver_state_publishing_thread_;
@@ -205,12 +275,91 @@ public:
    */
   void Start(
     const std::chrono::milliseconds motors_states_period,
-    const std::chrono::milliseconds driver_state_period);
+    const std::chrono::milliseconds driver_state_period)
+  {
+    canopen_communication_started_.store(false);
+    ctx_ = std::make_shared<lely::io::Context>();
+
+    canopen_communication_thread_ =
+      std::thread([this, motors_states_period, driver_state_period]() {
+        std::string slave_eds_path =
+          std::filesystem::path(
+            ament_index_cpp::get_package_share_directory("panther_hardware_interfaces")) /
+          "config" / "roboteq_motor_controllers_v80_21a.eds";
+        std::string slave1_eds_bin_path =
+          std::filesystem::path(
+            ament_index_cpp::get_package_share_directory("panther_hardware_interfaces")) /
+          "test" / "config" / "slave_1.bin";
+
+        std::string slave2_eds_bin_path =
+          std::filesystem::path(
+            ament_index_cpp::get_package_share_directory("panther_hardware_interfaces")) /
+          "test" / "config" / "slave_2.bin";
+
+        lely::io::IoGuard io_guard;
+        lely::io::Poll poll(*ctx_);
+        lely::ev::Loop loop(poll.get_poll());
+        auto exec = loop.get_executor();
+        lely::io::Timer timer(poll, exec, CLOCK_MONOTONIC);
+
+        lely::io::CanController ctrl("panther_can");
+
+        lely::io::CanChannel chan1(poll, exec);
+        chan1.open(ctrl);
+        lely::io::Timer timer1(poll, exec, CLOCK_MONOTONIC);
+        front_driver_ = std::make_shared<RoboteqSlave>(
+          timer1, chan1, slave_eds_path, slave1_eds_bin_path, 1);
+
+        lely::io::CanChannel chan2(poll, exec);
+        chan2.open(ctrl);
+        lely::io::Timer timer2(poll, exec, CLOCK_MONOTONIC);
+        rear_driver_ = std::make_shared<RoboteqSlave>(
+          timer2, chan2, slave_eds_path, slave2_eds_bin_path, 2);
+
+        front_driver_->Reset();
+        rear_driver_->Reset();
+        front_driver_->InitializeValues();
+        rear_driver_->InitializeValues();
+        front_driver_->StartPublishing(motors_states_period, driver_state_period);
+        rear_driver_->StartPublishing(motors_states_period, driver_state_period);
+
+        {
+          std::lock_guard<std::mutex> lck_g(canopen_communication_started_mtx_);
+          canopen_communication_started_.store(true);
+        }
+        canopen_communication_started_cond_.notify_all();
+
+        loop.run();
+
+        front_driver_->StopPublishing();
+        rear_driver_->StopPublishing();
+      });
+
+    if (!canopen_communication_started_.load()) {
+      std::unique_lock lck(canopen_communication_started_mtx_);
+      canopen_communication_started_cond_.wait(lck);
+    }
+
+    if (!canopen_communication_started_.load()) {
+      throw std::runtime_error("CAN communication not initialized");
+    }
+  }
 
   /**
    * @brief Stops CAN communication and removes simulated Roboteqs
    */
-  void Stop();
+  void Stop()
+  {
+    ctx_->shutdown();
+    if (canopen_communication_thread_.joinable()) {
+      canopen_communication_thread_.join();
+    }
+
+    front_driver_.reset();
+    rear_driver_.reset();
+
+    canopen_communication_started_.store(false);
+  }
 
   std::shared_ptr<RoboteqSlave> GetFrontDriver() { return front_driver_; }
   std::shared_ptr<RoboteqSlave> GetRearDriver() { return rear_driver_; }
@@ -230,4 +379,4 @@ private:
 
 }  // namespace panther_hardware_interfaces_test
 
-#endif  // PANTHER_HARDWARE_INTERFACES_ROBOTEQ_MOCK_HPP_
+#endif  // PANTHER_HARDWARE_INTERFACES_TEST_UTILS_ROBOTEQ_MOCK_HPP_
